+++ src/test/java/nextstep/subway/auth/acceptance/AuthAcceptanceTest.java
@@ -4,7 +4,7 @@ import nextstep.subway.AcceptanceTest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-public class AuthAcceptanceTest extends AcceptanceTest {
+class AuthAcceptanceTest extends AcceptanceTest {
 
     @DisplayName("Bearer Auth")
     @Test
