+++ src/test/java/nextstep/subway/path/PathAcceptanceTest.java
@@ -5,5 +5,6 @@ import org.junit.jupiter.api.DisplayName;
 
 
 @DisplayName("지하철 경로 조회")
-public class PathAcceptanceTest extends AcceptanceTest {
+class PathAcceptanceTest extends AcceptanceTest {
+
 }
